refactor: migrate ai to new architecture
