+++ core/src/db/schema/index.ts
@@ -1,5 +1,5 @@
-export * from "@/zap/db/schema/ai.sql";
-export * from "@/zap/db/schema/auth.sql";
-export * from "@/zap/db/schema/feedbacks.sql";
-export * from "@/zap/db/schema/notifications.sql";
-export * from "@/zap/db/schema/waitlist.sql";
+export * from "@/zap/ai/db/schema";
+export * from "@/zap-old/db/schema/auth.sql";
+export * from "@/zap-old/db/schema/feedbacks.sql";
+export * from "@/zap-old/db/schema/notifications.sql";
+export * from "@/zap-old/db/schema/waitlist.sql";
