+++ core/src/zap-old/lib/api/hooks/use-zap-query.ts
@@ -8,7 +8,7 @@ import {
 } from "@tanstack/react-query";
 import { useEffect, useMemo, useRef } from "react";
 
-import { handleClientError, handleSuccess } from "@/zap/lib/api/client";
+import { handleClientError, handleSuccess } from "@/zap-old/lib/api/client";
 
 export interface ZapQueryOptions<
   TQueryFnData,
