--- core/src/zap-old/services/ai/save-ai-settings.service.ts
@@ -1,27 +1,0 @@
-import "server-only";
-
-import type { AIProviderId, ModelName } from "@/zap/types/ai.types";
-
-import { saveOrUpdateAISettingsService } from "./save-or-update-ai-settings.service";
-
-interface SaveAISettingsService {
-  userId: string;
-  provider: AIProviderId;
-  model: ModelName;
-  apiKey: string;
-}
-
-export async function saveAISettingsService({
-  userId,
-  provider,
-  model,
-  apiKey,
-}: SaveAISettingsService) {
-  return await saveOrUpdateAISettingsService({
-    userId,
-    provider,
-    model,
-    apiKey,
-    mode: "create-only",
-  });
-}
