+++ core/src/zap/ai/components/api-key-input.tsx
