+++ core/src/zap/ai/components/model-select.tsx
