+++ core/src/zap/ai/components/provider-select.tsx
