+++ core/src/zap/ai/data/index.ts
@@ -1,4 +1,4 @@
-import type { AIProvider, AIProviderId, ModelName } from "@/zap/types/ai.types";
+import type { AIProvider, AIProviderId, ModelName } from "@/zap/ai/types";
 
 export const DEFAULT_MODEL: Record<AIProviderId, ModelName> = {
   openai: "gpt-4o-mini",
